Schemas seperated from endpoints to well formatted code files
