+++ backend/routers/schemas/__init__.py
@@ -1,0 +1,9 @@
+from .auth import (
+    AuthErrorResponse,
+    LoginRequest,
+    LoginSuccessResponse,
+    LogoutSuccessResponse,
+    RegisterSuccessResponse,
+    RegistrationRequest,
+    TokenRefreshSuccessResponse,
+)
